improved some code and added some handler structs
