+++ nakadion/examples/consume_subscription.rs
@@ -24,17 +24,18 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         .subscription_id(subscription_id)
         .inactivity_timeout_secs(30)
         .commit_timeout_millis(1500)
-        .update_stream_parameters(|p| p.batch_limit(10).max_uncommitted_events(1000))
+        .configure_stream_parameters(|p| p.batch_limit(10).max_uncommitted_events(1000))
         .connect_stream_timeout_secs(5);
 
+    // This is not necessary and just used
+    // to print the configured values
     builder.apply_defaults();
-
     println!("{:#?}", builder);
 
     let consumer =
         builder.build_with(client, handler::MyHandlerFactory, SlogLogger::new(logger))?;
 
-    let (handle, task) = consumer.start();
+    let (_handle, task) = consumer.start();
 
     let outcome = task.await;
 
@@ -55,7 +56,7 @@ fn main() {
 mod handler {
     use futures::future::{BoxFuture, FutureExt};
 
-    use nakadion::event_handler::*;
+    use nakadion::handler::*;
 
     pub struct MyHandler {
         count: usize,
@@ -90,7 +91,7 @@ mod handler {
 
         fn handler(
             &self,
-            assignment: &HandlerAssignment,
+            _assignment: &HandlerAssignment,
         ) -> BoxFuture<Result<Self::Handler, Error>> {
             async { Ok(MyHandler { count: 0 }) }.boxed()
         }
