+++ nakadion/src/lib.rs
@@ -9,7 +9,7 @@ pub(crate) use nakadi_types::Error;
 pub mod api;
 pub mod auth;
 pub mod consumer;
-pub mod event_handler;
+pub mod handler;
 
 pub(crate) mod event_stream;
 pub(crate) mod internals;
